[ADD] last test for offset

+++ src/modular_live_composer.c
@@ -632,7 +632,7 @@ void get_music_data(t_sensors *sensors)
 	printf("addr of 3: %d|", &sensors->time);
 
 	printf("Offsetof time : %d", offsetof(t_sensors, time));
-	printf("addr of 3: %d|", (uint32_t)sensors + (uint32_t)offsetof(t_sensors, time));
+	printf("addr of 3: %d|", (uint32_t)sensors + offsetof(t_sensors, time));
 
 	for (uint16_t i = 0; i < sensor_number; i++)
 	{
